tools: libboard: move print_eth_interface_info() to board_common
