+++ tools/libboard/board_common/board_common.c
@@ -36,6 +36,7 @@
 #include <fcntl.h>
 #include <sys/types.h>
 #include <sys/stat.h>
+#include <net/if.h>
 #include <net/ethernet.h>
 #include <opae/properties.h>
 #include <opae/utils.h>
@@ -52,6 +53,11 @@
 #define DFL_SYSFS_SEC_PR_ROOT                  DFL_SYSFS_SEC_GLOB "pr_root_entry_hash"
 #define DFL_SYSFS_SEC_SR_CANCEL                DFL_SYSFS_SEC_GLOB "sr_canceled_csks"
 #define DFL_SYSFS_SEC_SR_ROOT                  DFL_SYSFS_SEC_GLOB "sr_root_entry_hash"
+
+#define DFL_SYSFS_ETHINTERFACE   "dfl*.*/net/%s*"
+#define ETHTOOL_STR              "ethtool"
+#define IFCONFIG_STR             "ifconfig"
+#define IFCONFIG_UP_STR          "up"
 
 // Read sysfs
 fpga_result read_sysfs(fpga_token token, char *sysfs_path,
@@ -238,4 +244,83 @@ fpga_result print_sec_common_info(fpga_token token)
 	printf("********** SEC Info END ************ \n");
 
 	return resval;
+}
+
+// prints FPGA ethernet interface info
+fpga_result print_eth_interface_info(fpga_token token, const char *interface_name)
+{
+	fpga_result res                = FPGA_NOT_FOUND;
+	struct if_nameindex *if_nidxs  = NULL;
+	struct if_nameindex *intf      = NULL;
+	char cmd[SYSFS_PATH_MAX]       = { 0 };
+	char glob[SYSFS_PATH_MAX]      = { 0 };
+	int result                     = 0;
+	fpga_object fpga_object;
+
+	if_nidxs = if_nameindex();
+	if (if_nidxs != NULL) {
+		for (intf = if_nidxs; intf->if_index != 0
+			|| intf->if_name != NULL; intf++) {
+
+			char *p = strstr(intf->if_name, interface_name);
+			if (p) {
+				memset(glob, 0, sizeof(glob));
+				if (snprintf(glob, sizeof(glob),
+						DFL_SYSFS_ETHINTERFACE,
+						interface_name) < 0) {
+					OPAE_ERR("snprintf failed");
+					res = FPGA_EXCEPTION;
+					goto out_free;
+				}
+
+				// Check interface associated to bdf
+				res = fpgaTokenGetObject(token, glob,
+					&fpga_object, FPGA_OBJECT_GLOB);
+				if (res != FPGA_OK) {
+					OPAE_ERR("Failed to get token Object");
+					continue;
+				}
+				res = fpgaDestroyObject(&fpga_object);
+				if (res != FPGA_OK) {
+					OPAE_ERR("Failed to Destroy Object");
+				}
+
+				// Interface up
+				memset(cmd, 0, sizeof(cmd));
+				if (snprintf(cmd, sizeof(cmd),
+					"%s %s %s", IFCONFIG_STR, intf->if_name,
+					IFCONFIG_UP_STR) < 0) {
+					OPAE_ERR("snprintf failed");
+					res = FPGA_EXCEPTION;
+					goto out_free;
+				}
+				result = system(cmd);
+				if (result < 0) {
+					res = FPGA_EXCEPTION;
+					OPAE_ERR("Failed to run cmd: %s  %s",
+						cmd, strerror(errno));
+				}
+				// eth tool command
+				memset(cmd, 0, sizeof(cmd));
+				if (snprintf(cmd, sizeof(cmd),
+					"%s %s", ETHTOOL_STR, intf->if_name) < 0) {
+					OPAE_ERR("snprintf failed");
+					res = FPGA_EXCEPTION;
+					goto out_free;
+				}
+				result = system(cmd);
+				if (result < 0) {
+					res = FPGA_EXCEPTION;
+					OPAE_ERR("Failed to run cmd: %s  %s", cmd,
+						strerror(errno));
+				}
+
+			}
+		}
+
+out_free:
+		if_freenameindex(if_nidxs);
+	}
+
+	return res;
 }
