+++ tools/libboard/board_n3000/board_n3000.c
@@ -36,14 +36,11 @@
 #include <fcntl.h>
 #include <sys/types.h>
 #include <sys/stat.h>
-#include <net/ethernet.h>
 #include <opae/properties.h>
 #include <opae/utils.h>
 #include <opae/fpga.h>
 #include <sys/ioctl.h>
 #include <sys/types.h>
-#include <ifaddrs.h>
-#include <net/if.h>
 #include <stdlib.h>
 
 #include "../board_common/board_common.h"
@@ -72,13 +69,6 @@
 #define FPGA_PHYGROUP_MODE_4_25G              1
 #define FPGA_PHYGROUP_MODE_6_25G              3
 #define FPGA_PHYGROUP_MODE_2_2_25G            4
-
-#define ETHTOOL_STR              "ethtool"
-#define IFCONFIG_STR             "ifconfig"
-#define IFCONFIG_UP_STR          "up"
-#define FPGA_ETHINTERFACE_NAME   "npac"
-#define DFL_ETHINTERFACE         "dfl*.*/net/npac*"
-
 
 #define FPGA_BSID_REVISION(id)	(((id) >> 36) & 0xfff)
 #define FPGA_BSID_INTERFACE(id)	(((id) >> 32) & 0xf)
@@ -323,75 +313,6 @@ fpga_result print_board_info(fpga_token token)
 	return res;
 }
 
-// prints FPGA ethernet interface info
-fpga_result print_eth_interface_info(fpga_token token)
-{
-	fpga_result res                = FPGA_NOT_FOUND;
-	struct if_nameindex *if_nidxs  = NULL;
-	struct if_nameindex *intf      = NULL;
-	char cmd[SYSFS_PATH_MAX]       = { 0 };
-	int result                     = 0;
-	fpga_object fpga_object;
-
-	if_nidxs = if_nameindex();
-	if (if_nidxs != NULL) {
-		for (intf = if_nidxs; intf->if_index != 0
-			|| intf->if_name != NULL; intf++) {
-
-			char *p = strstr(intf->if_name, FPGA_ETHINTERFACE_NAME);
-			if (p) {
-				// Check interface associated to bdf
-				res = fpgaTokenGetObject(token, DFL_ETHINTERFACE,
-					&fpga_object, FPGA_OBJECT_GLOB);
-				if (res != FPGA_OK) {
-					OPAE_ERR("Failed to get token Object");
-					continue;
-				}
-				res = fpgaDestroyObject(&fpga_object);
-				if (res != FPGA_OK) {
-					OPAE_ERR("Failed to Destroy Object");
-				}
-
-				// Interface up
-				memset(cmd, 0, sizeof(cmd));
-				if (snprintf(cmd, sizeof(cmd),
-					"%s %s %s", IFCONFIG_STR, intf->if_name,
-					IFCONFIG_UP_STR) < 0) {
-					OPAE_ERR("snprintf failed");
-					res = FPGA_EXCEPTION;
-					goto out_free;
-				}
-				result = system(cmd);
-				if (result < 0) {
-					res = FPGA_EXCEPTION;
-					OPAE_ERR("Failed to run cmd: %s  %s",
-						cmd, strerror(errno));
-				}
-				// eth tool command
-				memset(cmd, 0, sizeof(cmd));
-				if (snprintf(cmd, sizeof(cmd),
-					"%s %s", ETHTOOL_STR, intf->if_name) < 0) {
-					OPAE_ERR("snprintf failed");
-					res = FPGA_EXCEPTION;
-					goto out_free;
-				}
-				result = system(cmd);
-				if (result < 0) {
-					res = FPGA_EXCEPTION;
-					OPAE_ERR("Failed to run cmd: %s  %s", cmd,
-						strerror(errno));
-				}
-
-			}
-		}
-
-out_free:
-		if_freenameindex(if_nidxs);
-	}
-
-	return res;
-}
-
 // print phy group information
 fpga_result print_phy_info(fpga_token token)
 {
@@ -403,9 +324,9 @@ fpga_result print_phy_info(fpga_token token)
 		return res;
 	}
 
-	res = print_eth_interface_info(token);
+	res = print_eth_interface_info(token, "npac");
 	if (res != FPGA_OK) {
-		OPAE_ERR("Failed to read phy group count");
+		OPAE_ERR("Failed to read phy info");
 		return res;
 	}
