+++ tools/libboard/board_n3000/board_n3000.h
@@ -200,15 +200,6 @@ fpga_result print_sec_info(fpga_token token);
 fpga_result print_sec_info(fpga_token token);
 
 /**
-* Prints ethernet interface info
-*
-* @param[in] token            fpga_token object for device (FPGA_DEVICE type)
-* @returns FPGA_OK on success. FPGA_EXCEPTION if eth interface not found.
-* FPGA_INVALID_PARAM if invalid parameters were provided
-*/
-fpga_result print_eth_interface_info(fpga_token token);
-
-/**
 * Prints fme verbose info
 *
 * @param[in] token            fpga_token object for device (FPGA_DEVICE type)
